[feat & refactor] : 폴더구조 개선 및 home 페이지 레이아웃 구성
1. home 페이지 레이아웃 구성
    - 로드맵, 투두 리스트 등을 composite 컴포넌트로 분리

2. composite 폴더구조 추가
    - feature, shared 이 컴포넌트들을 조합해 컴포넌트 구성
    - app 내의 페이지 이름과 composite 폴더 이름과 동일하게 가져감
    - 페이지 컴포넌트는 composite, features, shared 를 조합해서 구성

3. 폴더구조 개선 및 README 작성
    - 각 컴포넌트별로 component.ts, hooks.ts, api.ts, type.ts 파일 구성
    - 컴포넌트 중심의 모듈화 구조로 개선
