+++ src/composite/create-goal/confimGoalBottomBar/api.ts
@@ -1,0 +1,8 @@
+import { apiClient } from '@/shared/lib/apiClient';
+import { GoalFormData } from '@/shared/type/form';
+
+interface CreateGoalRequest extends GoalFormData {}
+
+export async function postCreateGoal(req: CreateGoalRequest) {
+  return await apiClient.post<CreateGoalRequest>('/goals', req);
+}
