+++ src/composite/create-goal/index.ts
@@ -1,0 +1,2 @@
+export { CreateGoalForm } from './createGoalForm/component';
+export { ConfirmGoalBottomBar } from './confimGoalBottomBar/component';
