+++ src/composite/home/cheerMessageCard/component.tsx
@@ -1,0 +1,7 @@
+export const CheerMessageCard = () => {
+  return (
+    <div className="flex bg-accent-fg-blue">
+      <p>히히 위로해줘용</p>
+    </div>
+  );
+};
