+++ src/composite/home/contributionGraph/component.tsx
@@ -1,0 +1,8 @@
+export const ContributionGraph = () => {
+  return (
+    <div className="flex flex-col w-full bg-amber-200">
+      <p>잔디 그래프</p>
+      <p>금잔디 아는사람?</p>
+    </div>
+  );
+};
