+++ src/composite/home/goalRoadMap/component.tsx
@@ -1,0 +1,8 @@
+export const GoalRoadMap = () => {
+  return (
+    <div className="flex flex-col w-full bg-accent-fg-orange">
+      <p>RoadMap</p>
+      <p>내 인생도 로드맵대로 갔으면</p>
+    </div>
+  );
+};
