+++ src/composite/home/index.ts
@@ -1,0 +1,5 @@
+export { TodayMissionBoard } from './todayMissionBoard/component';
+export { CheerMessageCard } from './cheerMessageCard/component';
+export { WeeklyPlanBoard } from './weeklyPlanBoard/component';
+export { GoalRoadMap } from './goalRoadMap/component';
+export { ContributionGraph } from './contributionGraph/component';
