+++ src/composite/home/todayMissionBoard/component.tsx
@@ -1,0 +1,7 @@
+export const TodayMissionBoard = () => {
+  return (
+    <div className="flex flex-col bg-accent-cyan">
+      <p>Today Mission</p>
+    </div>
+  );
+};
