+++ src/composite/login/loginForm/index.ts
@@ -1,0 +1,1 @@
+export { LoginForm } from './component';
