+++ src/composite/signup/signUpForm/index.ts
@@ -1,0 +1,1 @@
+export { SignUpForm } from './component';
