+++ src/feature/auth/index.ts
@@ -1,2 +1,3 @@
-export { LoginForm } from './components/LoginForm';
-export { LogoutButton } from './components/LogoutButton';
+export { LogoutButton } from './logoutButton/component';
+export { SelectJobButtonGroup } from './selectJobButtonGroup/component';
+export { SignupDialogButton } from './signupDialogButton/component';
