+++ src/feature/auth/logoutButton/component.tsx
