+++ src/feature/auth/selectJobButtonGroup/type.ts
@@ -1,0 +1,6 @@
+export type CareerYearType = 'NEWBIE' | 'JUNIOR' | 'MID' | 'SENIOR' | 'LEAD';
+
+export interface JobRole {
+  id: string;
+  name: string;
+}
