+++ src/feature/goal/index.ts
@@ -1,1 +1,3 @@
-export { ConfirmGoalBottomBar } from './components/ConfirmGoalBottomBar';
+export { ConfirmGoalDialogButton } from './confimGoalDialogButton/component';
+export { CreateGoalFormElement } from './createGoalFormElement/component';
+export { CreateGoalButton } from './createGoalButton/component';
