--- src/feature/goal/type.ts
@@ -1,12 +1,0 @@
-export interface CreateGoalFormData {
-  name: string;
-  duration: {
-    startDate: string;
-    endDate: string;
-  };
-  beforeAfter: {
-    asIs: string;
-    toBe: string;
-  };
-  plans: { content: string }[];
-}
