--- src/feature/onBoarding/index.ts
@@ -1,1 +1,0 @@
-export { useGoOnboarding } from './hooks/useGoOnboarding';
