+++ src/shared/type/form.ts
@@ -1,0 +1,15 @@
+export interface GoalFormData {
+  name: string;
+  duration: {
+    startDate: string;
+    endDate: string;
+  };
+  beforeAfter: {
+    asIs: string;
+    toBe: string;
+  };
+  plans: {
+    content: string;
+    weekOfMonth: number;
+  }[];
+}
